fix: harden real hardware boot and ring3 transition

+++ kernel/main.c
@@ -294,7 +294,7 @@ void kernel_main(const rixuri_boot_info_t *boot){
 	 ps2_keyboard_init();
 	 klog_write("BOOT: ps2 init done\r\n");
 	 klog_write("BOOT: ioapic init begin\r\n");
-	 int io_ready=0;if(acpi_ioapic_count()&&ioapic_init()==0){if(ioapic_route_irq(0,32,(uint8_t)lapic_id())!=0)panic("failed to route PIT IRQ");if(ioapic_route_irq(1,33,(uint8_t)lapic_id())!=0)klog_write("IOAPIC: failed to route IRQ1 (keyboard)\r\n");else{ioapic_unmask_irq(0);ioapic_unmask_irq(1);}pic_disable();io_ready=1;}
+	 int io_ready=0;if(acpi_ioapic_count()&&ioapic_init()==0){if(ioapic_route_irq(0,32,lapic_id())!=0)panic("failed to route PIT IRQ");if(ioapic_route_irq(1,33,lapic_id())!=0)klog_write("IOAPIC: failed to route IRQ1 (keyboard)\r\n");else{ioapic_unmask_irq(0);ioapic_unmask_irq(1);}pic_disable();io_ready=1;}
 	 klog_write("BOOT: ioapic init done\r\n");
  if(io_ready){idt_enable();klog_write("IRQ: PIT routed through IOAPIC; interrupts enabled\r\n");}
  else if(pic_init()==0){lapic_enable_pic_extint();idt_enable();klog_write("IRQ: IOAPIC unavailable; LAPIC ExtINT/PIC fallback enabled\r\n");}
